Implemented Shred Chamfer Similarity Comparision
Test also implemented in which we take two shreds, on which aligns
perfectly at 100 pixel offset.

Changed the THRESHOLD to discount the first and last element because
those are implicit features (in terms of convolution) and distort
the thresholding results.

Implemented more visualization helpers in the shred class

+++ Caruso/Caruso/Chamfer.cs
@@ -47,7 +47,7 @@ namespace Caruso
                 return chamfers;
             }
 
-            public static double[] ScanSimilarity(double[] chamfer1, double[] chamfer2)
+            public static double[] ScanSimilarity(int[] chamfer1, int[] chamfer2)
             {
                 var size1 = chamfer1.Length;
                 var size2 = chamfer2.Length;
@@ -76,7 +76,7 @@ namespace Caruso
                 return total;
             }
 
-            public static double Similarity(double[] smaller, double[] larger, int start=0)
+            public static double Similarity(int[] smaller, int[] larger, int start=0)
             {
                 if ( smaller.Length > larger.Length)
                 {
